+++ LibCarla/source/carla/opendrive/parser/GeoReferenceParser.cpp
@@ -18,14 +18,19 @@
 #include <string>
 #include <vector>
 #include <regex>
+#include <utility>
+#include <cctype>
+#include <stdexcept>
 
 namespace carla {
 namespace opendrive {
 namespace parser {
 
-  static std::string ToLowerCase(std::string s) {
-    std::transform(s.begin(), s.end(), s.begin(),
-                   [](unsigned char c){ return char(std::tolower(c)); });
+  template <typename S>
+  static std::string ToLowerCase(S&& s) {
+    std::string r = s;
+    for (auto& c : r)
+      c = static_cast<char>(std::tolower(c));
     return s;
   }
 
@@ -45,37 +50,132 @@ namespace parser {
     return geo_parameters_map;
   }
 
-  static geom::Ellipsoid CreateEllipsoid(std::unordered_map<std::string, std::string> parameters){
-    geom::Ellipsoid ellps;
+  static std::unordered_map<std::string, double> ParseOffsetParameters(const pugi::xml_node &geo_offset){
 
-    if (parameters.find("ellps") != parameters.end()) {
-      auto value = ToLowerCase(parameters["ellps"]);
+    std::unordered_map<std::string, double> geo_offset_result;
+
+    for (auto attr = geo_offset.first_attribute(); attr; attr = attr.next_attribute()){
+      std::string key = attr.name();
+      double value = attr.as_double();
+      geo_offset_result[key] = value;
+    }
+    return geo_offset_result;
+  }
+
+
+  template <typename M, typename S>
+  static bool TryGetParameter(std::string& out, M& parameters, S&& name)
+  {
+    auto i = parameters.find(std::forward<S>(name));
+    if (i == parameters.cend())
+      return false;
+    out = i->second;
+    return true;
+  }
+
+  template <typename M, typename S, typename O>
+  static
+  std::enable_if_t<std::is_floating_point<O>::value, bool>
+  TryGetParameter(O& out, M& parameters, S&& name)
+  {
+    auto i = parameters.find(std::forward<S>(name));
+    if (i == parameters.cend())
+      return false;
+    out = static_cast<O>(std::stod(i->second));
+    return true;
+  }
+
+  template <typename M, typename S, typename O>
+  static
+  std::enable_if_t<std::is_integral<O>::value && std::is_unsigned<O>::value, bool>
+  TryGetParameter(O& out, M& parameters, S&& name)
+  {
+    auto i = parameters.find(std::forward<S>(name));
+    if (i == parameters.cend())
+      return false;
+    out = static_cast<O>(std::stoull(i->second));
+    return true;
+  }
+
+  template <typename M, typename S, typename O>
+  static
+  std::enable_if_t<std::is_integral<O>::value && std::is_signed<O>::value, bool>
+  TryGetParameter(O& out, M& parameters, S&& name)
+  {
+    auto i = parameters.find(std::forward<S>(name));
+    if (i == parameters.cend())
+      return false;
+    out = static_cast<O>(std::stoll(i->second));
+    return true;
+  }
+
+  template <typename S, typename O>
+  static bool TryGetParameter(O& out,std::unordered_map<std::string, double>& parameters,S&& name)
+  {
+    auto i = parameters.find(std::forward<S>(name));
+    if (i == parameters.cend()){
+      return false;
+    }
+    out = static_cast<O>(i->second);
+    return true;
+  }
+
+
+  static geom::Ellipsoid CreateEllipsoid(std::unordered_map<std::string, std::string> parameters){
+
+    geom::Ellipsoid ellps;
+    std::string value;
+    bool initialized=false;
+    double x;
+
+    if (TryGetParameter(value, parameters, "ellps") || TryGetParameter(value, parameters, "datum")) {
+      value = ToLowerCase(value);
       auto it = geom::custom_ellipsoids.find(value);
       if (it != geom::custom_ellipsoids.end()) {
         ellps.a = it->second.first;
         ellps.f_inv = it->second.second;
-      } else {
-        auto val = geom::custom_ellipsoids.find("wgs84");
-        ellps.a = val->second.first;
-        ellps.f_inv = val->second.second;
+        initialized=true;
       }
     }
 
+    initialized |= TryGetParameter(ellps.a, parameters, "a");
     // Specific semi-major axis
-    if (parameters.find("a") != parameters.end()) {
-      ellps.a = std::stod(parameters["a"]);
+    if (TryGetParameter(x, parameters, "b")) {
+      initialized = true;
+      ellps.fromb(x);
     }
+    if (TryGetParameter(x, parameters, "f")) {
+      initialized = true;
+      ellps.fromf(x);
+    }
+    initialized |= TryGetParameter(ellps.f_inv, parameters, "rf");
 
-    // Specific semi-minor axis
-    if (parameters.find("b") != parameters.end()) {
-      ellps.fromb(std::stod(parameters["b"]));
-    } else if (parameters.find("f") != parameters.end()) {
-      ellps.fromf(std::stod(parameters["f"]));
-    } else if (parameters.find("rf") != parameters.end()) {
-      ellps.f_inv = std::stod(parameters["rf"]);
+    if (!initialized)    {
+      auto val = geom::custom_ellipsoids.find("wgs84");
+      ellps.a = val->second.first;
+      ellps.f_inv = val->second.second;
     }
 
     return ellps;
+  }
+
+  static boost::optional<geom::OffsetTransform> CreateOffsetTransform(std::unordered_map<std::string, double>& offset_parameters){
+
+    if(offset_parameters.empty()){
+      return boost::none;
+    }
+
+    double x = 0.0;
+    double y = 0.0;
+    double z = 0.0;
+    double hdg = 0.0;
+
+    TryGetParameter(x, offset_parameters, "x");
+    TryGetParameter(y, offset_parameters, "y");
+    TryGetParameter(z, offset_parameters, "z");
+    TryGetParameter(hdg, offset_parameters, "hdg");
+
+    return geom::OffsetTransform{x, y, z, hdg};
   }
 
   static geom::GeoProjection CreateTransverseMercatorProjection(
@@ -84,21 +184,11 @@ namespace parser {
     geom::Ellipsoid ellipsoid){
 
     geom::TransverseMercatorParams p;
-    if (parameters.find("lat_0") != parameters.end()) {
-      p.lat_0 = std::stod(parameters["lat_0"]);
-    }
-    if (parameters.find("lon_0") != parameters.end()) {
-      p.lon_0 = std::stod(parameters["lon_0"]);
-    }
-    if (parameters.find("k") != parameters.end()) {
-      p.k = std::stod(parameters["k"]);
-    }
-    if (parameters.find("lat_0") != parameters.end()) {
-      p.x_0 = std::stod(parameters["x_0"]);
-    }
-    if (parameters.find("lat_0") != parameters.end()) {
-      p.y_0 = std::stod(parameters["y_0"]);
-    }
+    TryGetParameter(p.lat_0, parameters, "lat_0");
+    TryGetParameter(p.lon_0, parameters, "lon_0");
+    TryGetParameter(p.k, parameters, "k");
+    TryGetParameter(p.x_0, parameters, "x_0");
+    TryGetParameter(p.y_0, parameters, "y_0");
     p.ellps = ellipsoid;
     auto projection = geom::GeoProjection::Make(p);
     projection.setPROJString(proj_string);
@@ -108,16 +198,15 @@ namespace parser {
   static geom::GeoProjection CreateUniversalTransverseMercatorProjection(
     std::unordered_map<std::string, std::string> parameters,
     std::string proj_string,
-    geom::Ellipsoid ellipsoid){
+    geom::Ellipsoid ellipsoid, boost::optional<carla::geom::OffsetTransform> offset){
 
     geom::UniversalTransverseMercatorParams p;
-    if (parameters.find("zone") != parameters.end()) {
-      p.zone = std::stoi(parameters["zone"]);
-    } else {
+    if (!TryGetParameter(p.zone, parameters, "zone")) {
       log_warning("Missing 'zone' parameter for UTM projection. Using default value " + std::to_string(p.zone));
     }
     p.north = (parameters.count("south") > 0) ? false : true;
     p.ellps = ellipsoid;
+    p.offset = offset;
     auto projection = geom::GeoProjection::Make(p);
     projection.setPROJString(proj_string);
     return projection;
@@ -141,30 +230,14 @@ namespace parser {
     geom::Ellipsoid ellipsoid){
 
     geom::LambertConformalConicParams p;
-
-    if (parameters.find("lon_0") != parameters.end()) {
-      p.lon_0 = std::stod(parameters["lon_0"]);
-    }
-    if (parameters.find("lat_0") != parameters.end()) {
-      p.lat_0 = std::stod(parameters["lat_0"]);
-    }
-    if (parameters.find("lat_1") != parameters.end()) {
-      p.lat_1 = std::stod(parameters["lat_1"]);
-    } else {
+    TryGetParameter(p.lon_0, parameters, "lon_0");
+    TryGetParameter(p.lat_0, parameters, "lat_0");
+    if (!TryGetParameter(p.lat_1, parameters, "lat_1"))
       log_warning("Missing 'lat_1' parameter for LCC projection. Using default value " + std::to_string(p.lat_1));
-    }
-    if (parameters.find("lat_2") != parameters.end()) {
-      p.lat_2 = std::stod(parameters["lat_2"]);
-    } else {
+    if (!TryGetParameter(p.lat_2, parameters, "lat_2"))
       log_warning("Missing 'lat_2' parameter for LCC projection. Using default value " + std::to_string(p.lat_2));
-    }
-    if (parameters.find("x_0") != parameters.end()) {
-      p.x_0 = std::stod(parameters["x_0"]);
-    }
-    if (parameters.find("y_0") != parameters.end()) {
-      p.y_0 = std::stod(parameters["y_0"]);
-    }
-
+    TryGetParameter(p.x_0, parameters, "x_0");
+    TryGetParameter(p.y_0, parameters, "y_0");
     p.ellps = ellipsoid;
     auto projection = geom::GeoProjection::Make(p);
     projection.setPROJString(proj_string);
@@ -182,23 +255,16 @@ namespace parser {
 
   static geom::GeoLocation CreateTransverseMercatorGeoReference(std::unordered_map<std::string, std::string> parameters){
     geom::GeoLocation result{0.0, 0.0, 0.0};
-    if (parameters.find("lat_0") != parameters.end()) {
-      result.latitude = std::stod(parameters["lat_0"]);
-    }
-    if (parameters.find("lon_0") != parameters.end()) {
-      result.longitude = std::stod(parameters["lon_0"]);
-    }
+    TryGetParameter(result.latitude, parameters, "lat_0");
+-   TryGetParameter(result.longitude, parameters, "lon_0");
     return result;
   }
 
   static geom::GeoLocation CreateUniversalTransverseMercatorGeoReference(std::unordered_map<std::string, std::string> parameters){
     geom::GeoLocation result{0.0, 0.0, 0.0};
-
     result.latitude = 0.0;
-    if (parameters.find("zone") != parameters.end()) {
-      double zone = std::stod(parameters["zone"]);
-      result.longitude = geom::Math::ToRadians(6 * zone - 183);
-    }
+    if (TryGetParameter(result.longitude, parameters, "zone"))
+      result.longitude = geom::Math::ToRadians(6 * result.longitude - 183);
     return result;
   }
 
@@ -209,12 +275,8 @@ namespace parser {
 
   static geom::GeoLocation CreateLamberConic2SPGeoReference(std::unordered_map<std::string, std::string> parameters){
     geom::GeoLocation result{0.0, 0.0, 0.0};
-    if (parameters.find("lat_0") != parameters.end()) {
-      result.latitude = std::stod(parameters["lat_0"]);
-    }
-    if (parameters.find("lon_0") != parameters.end()) {
-      result.longitude = std::stod(parameters["lon_0"]);
-    }
+    TryGetParameter(result.latitude, parameters, "lat_0");
+    TryGetParameter(result.longitude, parameters, "lon_0");
     return result;
   }
 
@@ -223,73 +285,68 @@ namespace parser {
     return result;
   }
 
-  static geom::GeoProjection ParseGeoProjection(const std::string &proj_string) {
-    geom::GeoProjection result;
-
+  static auto ParseGeoProjectionAndReference(const std::string& proj_string, const pugi::xml_node &geo_offset){
     auto parameters = ParseProjectionParameters(proj_string);
+    auto offset_parameters = ParseOffsetParameters(geo_offset);
     auto ellipsoid = CreateEllipsoid(parameters);
+    auto offset_transform = CreateOffsetTransform(offset_parameters);
 
     // Get the projection type
     std::string proj;
-    if (parameters.find("proj") != parameters.end()) {
-      proj = parameters["proj"];
-    } else {
+    if (!TryGetParameter(proj, parameters, "proj")) {
       log_warning("cannot find the type of projection, using default transverse mercator");
-      return CreateDefaultProjection(ellipsoid);
+      return std::make_pair(
+        CreateDefaultProjection(ellipsoid),
+        CreateDefaultGeoReference());
     }
 
     // Parse the parameters
     if (proj == "tmerc") {
-      return CreateTransverseMercatorProjection(parameters, proj_string, ellipsoid);
+      return std::make_pair(
+        CreateTransverseMercatorProjection(parameters, proj_string, ellipsoid),
+        CreateTransverseMercatorGeoReference(parameters));
     } else if (proj == "utm") {
-      return CreateUniversalTransverseMercatorProjection(parameters, proj_string, ellipsoid);
+      return std::make_pair(
+        CreateUniversalTransverseMercatorProjection(parameters, proj_string, ellipsoid, offset_transform),
+        CreateUniversalTransverseMercatorGeoReference(parameters));
     } else if (proj == "merc") {
-      // Fixed projection, doesn't have any parameters
-      return CreateWebMercatorProjection(proj_string, ellipsoid);
+      return std::make_pair(
+        CreateWebMercatorProjection(proj_string, ellipsoid),
+        CreateWebMercatorGeoReference());
     } else if (proj == "lcc") {
-      return CreateLambertConformalConicProjection(parameters, proj_string, ellipsoid);
-    }
-
-    log_debug("projection '" + proj + "' is not supported, using default transverse mercator.");
-    return CreateDefaultProjection(ellipsoid);
-  }
-
-  static geom::GeoLocation ParseGeoReference(const std::string &proj_string) {
-
-    auto parameters = ParseProjectionParameters(proj_string);
-
-    // Get the projection type
-    std::string proj;
-    if (parameters.find("proj") != parameters.end()) {
-      proj = parameters["proj"];
+      return std::make_pair(
+        CreateLambertConformalConicProjection(parameters, proj_string, ellipsoid),
+        CreateLamberConic2SPGeoReference(parameters));
     } else {
-      log_warning("cannot find the type of projection, using default geolocation");
-      return CreateDefaultGeoReference();
+      log_debug("projection '" + proj + "' is not supported, using default transverse mercator.");
+      return std::make_pair(
+        CreateDefaultProjection(ellipsoid),
+        CreateDefaultGeoReference());
     }
-
-    if (proj == "tmerc") {
-      return CreateTransverseMercatorGeoReference(parameters);
-    } else if (proj == "utm") {
-      return CreateUniversalTransverseMercatorGeoReference(parameters);
-    } else if (proj == "merc") {
-      // Fixed projection, doesn't have any parameters
-      return CreateWebMercatorGeoReference();
-    } else if (proj == "lcc") {
-      return CreateLamberConic2SPGeoReference(parameters);
-    }
-
-    log_debug("projection '" + proj + "' is not supported, using default geolocation.");
-    return CreateDefaultGeoReference();
   }
 
   void GeoReferenceParser::Parse(
       const pugi::xml_document &xml,
       carla::road::MapBuilder &map_builder) {
-    map_builder.SetGeoReference(ParseGeoReference(
-        xml.child("OpenDRIVE").child("header").child_value("geoReference")));
 
-    map_builder.SetGeoProjection(ParseGeoProjection(
-      xml.child("OpenDRIVE").child("header").child_value("geoReference")));
+    auto georeference_string =
+      xml
+      .child("OpenDRIVE")
+      .child("header")
+      .child_value("geoReference");
+
+    auto georeference_offset_node =
+      xml
+      .child("OpenDRIVE")
+      .child("header")
+      .child("offset");
+
+    auto reference_and_projection = ParseGeoProjectionAndReference(
+      georeference_string, georeference_offset_node);
+    map_builder.SetGeoProjection(reference_and_projection.first);
+    map_builder.SetGeoReference(reference_and_projection.second);
+
+
   }
 
 } // parser