+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Create class with functions to import points and polylines from satellite segmentation (#8946, #8949 #8950)
 * Added Visual C++ Redistributable prerequisite and troubleshooting note to Windows quickstart guide (#9560)
 * Added Docker-based development environment for CARLA UE5
+* Added support for parsing offsets from OpenDRIVE using optional offset transforms.
 
 ## CARLA 0.10.0